Refactor TT commands
- Removed cases were any was used as type
- Removed unnecessary tr-catch blocks
- Removed unnecessary triple equality operators

I hope everything is linted fine, I ran the lint-fix command
